feat: add tracking to correct location for platform webview desktop
