+++ apps/ledger-live-desktop/src/renderer/components/Web3AppWebview/PlatformAPIWebview.tsx
@@ -28,7 +28,6 @@ import { openModal } from "../../actions/modals";
 import { flattenAccountsSelector } from "../../reducers/accounts";
 import BigSpinner from "../BigSpinner";
 import { track } from "~/renderer/analytics/segment";
-import { currentRouteNameRef } from "~/renderer/analytics/screenRefs";
 import {
   requestAccountLogic,
   broadcastTransactionLogic,
@@ -38,18 +37,7 @@ import { Loader } from "./styled";
 import { WebviewAPI, WebviewProps } from "./types";
 import { useWebviewState } from "./helpers";
 
-const tracking = trackingWrapper(
-  (eventName: string, properties?: Record<string, unknown> | null, mandatory?: boolean | null) =>
-    track(
-      eventName,
-      {
-        ...properties,
-        flowInitiatedFrom:
-          currentRouteNameRef.current === "Platform Catalog" ? "Discover" : "Native",
-      },
-      mandatory,
-    ),
-);
+const tracking = trackingWrapper(track);
 
 export const PlatformAPIWebview = forwardRef<WebviewAPI, WebviewProps>(
   ({ manifest, inputs = {}, onStateChange }, ref) => {
